+++ app/src/main/java/org/simple/clinic/teleconsultlog/prescription/TeleconsultPrescriptionScreen.kt
@@ -5,6 +5,7 @@ import android.os.Parcelable
 import android.util.AttributeSet
 import androidx.constraintlayout.widget.ConstraintLayout
 import com.jakewharton.rxbinding3.appcompat.navigationClicks
+import com.jakewharton.rxbinding3.view.clicks
 import io.reactivex.Observable
 import io.reactivex.rxkotlin.ofType
 import kotlinx.android.synthetic.main.screen_teleconsult_prescription.view.*
@@ -40,7 +41,11 @@ class TeleconsultPrescriptionScreen constructor(
   }
 
   private val events by unsafeLazy {
-    backClicks()
+    Observable
+        .merge(
+            backClicks(),
+            nextClicks()
+        )
         .compose(ReportAnalyticsEvents())
   }
 
@@ -49,7 +54,7 @@ class TeleconsultPrescriptionScreen constructor(
 
     MobiusDelegate.forView(
         events = events.ofType(),
-        defaultModel = TeleconsultPrescriptionModel.create(screenKey.patientUuid),
+        defaultModel = TeleconsultPrescriptionModel.create(screenKey.teleconsultRecordId, screenKey.patientUuid),
         init = TeleconsultPrescriptionInit(),
         update = TeleconsultPrescriptionUpdate(),
         effectHandler = effectHandler.create(this).build(),
@@ -96,10 +101,28 @@ class TeleconsultPrescriptionScreen constructor(
     screenRouter.pop()
   }
 
+  override fun showSignatureRequiredError() {
+    teleconsultPrescriptionDoctorInfoView.showSignatureError()
+  }
+
   private fun backClicks(): Observable<UiEvent> {
     return toolbar
         .navigationClicks()
         .map { BackClicked }
+  }
+
+  private fun nextClicks(): Observable<UiEvent> {
+    return nextButton
+        .clicks()
+        .map {
+          val medicalInstructions = teleconsultPrescriptionDoctorInfoView.medicalInstructions
+          val medicalRegistrationId = teleconsultPrescriptionDoctorInfoView.medicalRegistrationId
+
+          NextButtonClicked(
+              medicalInstructions = medicalInstructions,
+              medicalRegistrationId = medicalRegistrationId
+          )
+        }
   }
 
   interface Injector {