+++ app/src/main/java/org/simple/clinic/teleconsultlog/prescription/TeleconsultPrescriptionUiActions.kt
@@ -2,4 +2,5 @@ package org.simple.clinic.teleconsultlog.prescription
 
 interface TeleconsultPrescriptionUiActions {
   fun goBackToPreviousScreen()
+  fun showSignatureRequiredError()
 }