+++ app/src/main/java/org/simple/clinic/teleconsultlog/prescription/TeleconsultPrescriptionUpdate.kt
@@ -1,6 +1,7 @@
 package org.simple.clinic.teleconsultlog.prescription
 
 import com.spotify.mobius.Next
+import com.spotify.mobius.Next.noChange
 import com.spotify.mobius.Update
 import org.simple.clinic.mobius.dispatch
 import org.simple.clinic.mobius.next
@@ -14,6 +15,20 @@ class TeleconsultPrescriptionUpdate : Update<TeleconsultPrescriptionModel, Telec
     return when (event) {
       is PatientDetailsLoaded -> next(model.patientLoaded(event.patient))
       BackClicked -> dispatch(GoBack)
+      is DataForNextClickLoaded -> dataForNextClickLoaded(event)
+      is NextButtonClicked -> dispatch(LoadDataForNextClick(
+          teleconsultRecordId = model.teleconsultRecordId,
+          medicalInstructions = event.medicalInstructions,
+          medicalRegistrationId = event.medicalRegistrationId
+      ))
+    }
+  }
+
+  private fun dataForNextClickLoaded(event: DataForNextClickLoaded): Next<TeleconsultPrescriptionModel, TeleconsultPrescriptionEffect> {
+    return if (event.signatureBitmap == null) {
+      dispatch(ShowSignatureRequiredError)
+    } else {
+      noChange()
     }
   }
 }