+++ app/src/main/java/org/simple/clinic/teleconsultlog/prescription/doctorinfo/TeleconsultDoctorInfoEffect.kt
@@ -13,3 +13,7 @@ object LoadSignatureBitmap : TeleconsultDoctorInfoEffect()
 data class SetSignatureBitmap(val bitmap: Bitmap) : TeleconsultDoctorInfoEffect()
 
 object LoadCurrentUser : TeleconsultDoctorInfoEffect()
+
+object ShowAddSignatureDialog : TeleconsultDoctorInfoEffect()
+
+object ShowAddSignatureButton : TeleconsultDoctorInfoEffect()