+++ app/src/test/java/org/simple/clinic/teleconsultlog/prescription/doctorinfo/TeleconsultDoctorInfoUpdateTest.kt
@@ -44,6 +44,17 @@ class TeleconsultDoctorInfoUpdateTest {
   }
 
   @Test
+  fun `when there is no signature bitmap, then show add signature button`() {
+    updateSpec
+        .given(model)
+        .whenEvent(SignatureBitmapLoaded(null))
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(ShowAddSignatureButton)
+        ))
+  }
+
+  @Test
   fun `when medical registration id changes, then update the model`() {
     val oldMedicalRegistrationId = "1234567890"
     val newMedicalRegistrationId = "0987654321"
@@ -84,6 +95,28 @@ class TeleconsultDoctorInfoUpdateTest {
         .then(assertThatNext(
             hasModel(model.currentUserLoaded(currentUser)),
             hasNoEffects()
+        ))
+  }
+
+  @Test
+  fun `when add signature button is clicked, then show the add signature dialog`() {
+    updateSpec
+        .given(model)
+        .whenEvent(AddSignatureClicked)
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(ShowAddSignatureDialog)
+        ))
+  }
+
+  @Test
+  fun `when activity is resumed, then load the signature`() {
+    updateSpec
+        .given(model)
+        .whenEvent(ActivityResumed)
+        .then(assertThatNext(
+            hasNoModel(),
+            hasEffects(LoadSignatureBitmap)
         ))
   }
 }