+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # CHANGELOG
 
 ## Next Release
+## Features
+- [In Progress: 22 Sep 2020] Add `TeleconsultPrescriptionScreen`
+
 ### Internal
 - Integrated Firebase Performance Monitoring
 - Remove `PrescribedDrugs` from `TeleconsultRecord`
@@ -10,7 +13,6 @@
 ### Features
 - Add `TeleconsultRecordScreen`
 - Show `TeleconsultLogNotAllowed` dialog when Medical officer is not allowed to teleconsult
-- [In Progress: 21 Sep 2020] Add `TeleconsultPrescriptionScreen`
 
 ### Internal
 - [In Progress: 27 Aug 2020] Add support for sending teleconsult request via SMS